Separate socket initialization from app index
